requirements.txt
tqdm
tensorflow>=1.4.0
regex
flask_socketio
eventlet
gevent
colorama
opencv-python
cmake
numpy
dlib
flask_sqlalchemy
flask_script
flask_migrate
redis
sklearn
scipy
passlib
Pillow
requests
faker
mysqlclient
boto3
Flask-Session